pyproject.toml
[tool.poetry]
name = "graphrag-on-law"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
numpy = {version = "^1.21.0", allow-prereleases = true}
scikit-learn = "0.24.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
